Migrate code to new file.
