+++ dronegen/misc.go
@@ -20,134 +20,6 @@ import (
 	"strings"
 )
 
-const (
-	// StagingRegistry is the staging registry images are pushed to before being promoted to the production registry.
-	StagingRegistry = "146628656107.dkr.ecr.us-west-2.amazonaws.com"
-
-	// ProductionRegistry is the production image registry that hosts are customer facing container images.
-	ProductionRegistry = "public.ecr.aws"
-
-	// ProductionRegistryQuay is the production image registry that hosts images on quay.io. Will be deprecated in the future.
-	// See RFD 73 - https://github.com/gravitational/teleport/blob/c18c09f5d562dd46a509154eab4295ad39decc3c/rfd/0073-public-image-registry.md
-	ProductionRegistryQuay = "quay.io"
-)
-
-func promoteBuildPipelines() []pipeline {
-	aptPipeline := promoteAptPipeline()
-	dockerPipelineECR := buildDockerPromotionPipelineECR()
-	dockerPipelineQuay := buildDockerPromotionPipelineQuay()
-	return []pipeline{aptPipeline, dockerPipelineECR, dockerPipelineQuay}
-}
-
-func buildDockerPromotionPipelineECR() pipeline {
-	dockerPipeline := newKubePipeline("promote-docker-ecr")
-	dockerPipeline.Trigger = triggerPromote
-	dockerPipeline.Trigger.Target.Include = append(dockerPipeline.Trigger.Target.Include, "promote-docker", "promote-docker-ecr")
-	dockerPipeline.Workspace = workspace{Path: "/go"}
-
-	// Add docker service
-	dockerPipeline.Services = []service{
-		dockerService(),
-	}
-	dockerPipeline.Volumes = dockerVolumes()
-
-	dockerPipeline.Steps = append(dockerPipeline.Steps, verifyTaggedBuildStep())
-	dockerPipeline.Steps = append(dockerPipeline.Steps, waitForDockerStep())
-
-	// Pull/Push Steps
-	dockerPipeline.Steps = append(dockerPipeline.Steps, step{
-		Name:  "Pull/retag Docker images",
-		Image: "docker",
-		Environment: map[string]value{
-			"AWS_ACCESS_KEY_ID":     {fromSecret: "PRODUCTION_TELEPORT_DRONE_USER_ECR_KEY"},
-			"AWS_SECRET_ACCESS_KEY": {fromSecret: "PRODUCTION_TELEPORT_DRONE_USER_ECR_SECRET"},
-		},
-		Volumes: dockerVolumeRefs(),
-		Commands: []string{
-			"apk add --no-cache aws-cli",
-			"export VERSION=${DRONE_TAG##v}",
-			// authenticate with staging credentials
-			"aws ecr get-login-password --region=us-west-2 | docker login -u=\"AWS\" --password-stdin " + StagingRegistry,
-			// pull staging images
-			"echo \"---> Pulling images for $${VERSION}\"",
-			fmt.Sprintf("docker pull %s/gravitational/teleport:$${VERSION}", StagingRegistry),
-			fmt.Sprintf("docker pull %s/gravitational/teleport-ent:$${VERSION}", StagingRegistry),
-			fmt.Sprintf("docker pull %s/gravitational/teleport-ent:$${VERSION}-fips", StagingRegistry),
-			// retag images to production naming
-			"echo \"---> Tagging images for $${VERSION}\"",
-			fmt.Sprintf("docker tag %s/gravitational/teleport:$${VERSION} %s/gravitational/teleport:$${VERSION}", StagingRegistry, ProductionRegistry),
-			fmt.Sprintf("docker tag %s/gravitational/teleport-ent:$${VERSION} %s/gravitational/teleport-ent:$${VERSION}", StagingRegistry, ProductionRegistry),
-			fmt.Sprintf("docker tag %s/gravitational/teleport-ent:$${VERSION}-fips %s/gravitational/teleport-ent:$${VERSION}-fips", StagingRegistry, ProductionRegistry),
-			// authenticate with production credentials
-			"docker logout " + StagingRegistry,
-			"aws ecr-public get-login-password --region=us-east-1 | docker login -u=\"AWS\" --password-stdin " + ProductionRegistry,
-			// push production images
-			"echo \"---> Pushing images for $${VERSION}\"",
-			// push production images ECR
-			fmt.Sprintf("docker push %s/gravitational/teleport:$${VERSION}", ProductionRegistry),
-			fmt.Sprintf("docker push %s/gravitational/teleport-ent:$${VERSION}", ProductionRegistry),
-			fmt.Sprintf("docker push %s/gravitational/teleport-ent:$${VERSION}-fips", ProductionRegistry),
-		},
-	})
-
-	return dockerPipeline
-}
-
-func buildDockerPromotionPipelineQuay() pipeline {
-	dockerPipeline := newKubePipeline("promote-docker-quay")
-	dockerPipeline.Trigger = triggerPromote
-	dockerPipeline.Trigger.Target.Include = append(dockerPipeline.Trigger.Target.Include, "promote-docker", "promote-docker-quay")
-	dockerPipeline.Workspace = workspace{Path: "/go"}
-
-	// Add docker service
-	dockerPipeline.Services = []service{
-		dockerService(),
-	}
-	dockerPipeline.Volumes = dockerVolumes()
-
-	dockerPipeline.Steps = append(dockerPipeline.Steps, verifyTaggedBuildStep())
-	dockerPipeline.Steps = append(dockerPipeline.Steps, waitForDockerStep())
-
-	// Pull/Push Steps
-	dockerPipeline.Steps = append(dockerPipeline.Steps, step{
-		Name:  "Pull/retag Docker images",
-		Image: "docker",
-		Environment: map[string]value{
-			"AWS_ACCESS_KEY_ID":     {fromSecret: "STAGING_TELEPORT_DRONE_USER_ECR_KEY"},
-			"AWS_SECRET_ACCESS_KEY": {fromSecret: "STAGING_TELEPORT_DRONE_USER_ECR_SECRET"},
-			"QUAY_USERNAME":         {fromSecret: "PRODUCTION_QUAYIO_DOCKER_USERNAME"},
-			"QUAY_PASSWORD":         {fromSecret: "PRODUCTION_QUAYIO_DOCKER_PASSWORD"},
-		},
-		Volumes: dockerVolumeRefs(),
-		Commands: []string{
-			"apk add --no-cache aws-cli",
-			"export VERSION=${DRONE_TAG##v}",
-			// authenticate with staging credentials
-			"aws ecr get-login-password --region=us-west-2 | docker login -u=\"AWS\" --password-stdin " + StagingRegistry,
-			// pull staging images
-			"echo \"---> Pulling images for $${VERSION}\"",
-			fmt.Sprintf("docker pull %s/gravitational/teleport:$${VERSION}", StagingRegistry),
-			fmt.Sprintf("docker pull %s/gravitational/teleport-ent:$${VERSION}", StagingRegistry),
-			fmt.Sprintf("docker pull %s/gravitational/teleport-ent:$${VERSION}-fips", StagingRegistry),
-			// retag images to production naming
-			"echo \"---> Tagging images for $${VERSION}\"",
-			fmt.Sprintf("docker tag %s/gravitational/teleport:$${VERSION} %s/gravitational/teleport:$${VERSION}", StagingRegistry, ProductionRegistryQuay),
-			fmt.Sprintf("docker tag %s/gravitational/teleport-ent:$${VERSION} %s/gravitational/teleport-ent:$${VERSION}", StagingRegistry, ProductionRegistryQuay),
-			fmt.Sprintf("docker tag %s/gravitational/teleport-ent:$${VERSION}-fips %s/gravitational/teleport-ent:$${VERSION}-fips", StagingRegistry, ProductionRegistryQuay),
-			// authenticate with production credentials
-			"docker logout " + StagingRegistry,
-			"docker login -u=\"$QUAY_USERNAME\" -p=\"$QUAY_PASSWORD\" " + ProductionRegistryQuay,
-			// push production images
-			"echo \"---> Pushing images for $${VERSION}\"",
-			fmt.Sprintf("docker push %s/gravitational/teleport:$${VERSION}", ProductionRegistryQuay),
-			fmt.Sprintf("docker push %s/gravitational/teleport-ent:$${VERSION}", ProductionRegistryQuay),
-			fmt.Sprintf("docker push %s/gravitational/teleport-ent:$${VERSION}-fips", ProductionRegistryQuay),
-		},
-	})
-
-	return dockerPipeline
-}
-
 // Used for one-off migrations of older versions.
 // Use cases include:
 //  * We want to support another OS while providing backwards compatibility
